requirements.txt
discord
python-dotenv
jishaku